:recycle: Simplify packages
