+++ example/client/src/main/scala/samples/EnumSample.scala
@@ -1,8 +1,6 @@
 package samples
 
-import dev.cheleb.scalamigen.*
-import dev.cheleb.scalamigen.forms.given
-import dev.cheleb.scalamigen.forms.*
+import dev.cheleb.scalamigen.{*, given}
 
 import com.raquo.laminar.api.L.*
